requirements.txt
google-generativeai
python-dotenv
openai
phidata
yfinance
packaging
duckduckgo-search
fastapi
groq